рефакторит: добавляет HOC PageTemplate
