+++ src/components/main-page/main-page.tsx
@@ -1,13 +1,10 @@
-import Header from '../header/header';
-import Footer from '../footer/footer';
 import { Link } from 'react-router-dom';
 import { AppRoute } from '../../const/app-route';
+import PageTemplate from '../../hocs/page-template';
 
 function MainPage(): JSX.Element {
   return(
-    <div className="wrapper">
-      <Header/>
-
+    <PageTemplate>
       <main className="page-content">
         <div className="container">
           <h1>Главная</h1>
@@ -16,9 +13,7 @@ function MainPage(): JSX.Element {
           </ol>
         </div>
       </main>
-
-      <Footer/>
-    </div>
+    </PageTemplate>
   );
 }
